requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
jproperties==2.1.1
kafka-python==2.0.2
MarkupSafe==2.0.1
nltk==3.6.7
numpy==1.22.2
pandas==1.4.0
pymongo==4.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
regex==2022.1.18
requests==2.27.1
scikit-learn==1.0.2
scikit-surprise==1.1.1
scipy==1.8.0
six==1.16.0
threadpoolctl==3.1.0
tqdm==4.62.3
urllib3==1.26.8
Werkzeug==2.0.3
pytest==7.1.1
pytest-cov==3.0.0
coverage==6.3.2
pylint==2.13.0
Flask-PyMongo==2.3.0
fluent-logger==0.10.0
neptune-client==0.15.2
psutil==5.9.0